dbauction/StreamLitApp/requirements.txt
mysqlclient==2.2.0
mysql-connector-python==8.2.0
PyMySQL==1.0.2
streamlit==1.28.2
openAI==0.28

